+++ CHANGES.rst
@@ -10,6 +10,10 @@ Changelog
   - Added `utils.get_annexes_config` function to be able to get the annexes config
     depending on `context` and annex `portal_type`.
     [gbastien]
+  - Added `the_objects=False` parameter to `ToolPloneMeeting.get_plone_groups_for_user`
+    to get `GroupData` instances instead group ids.
+    This is used by the `plonemeeting.restapi` `@users` endpoint.
+    [gbastien]
 
 4.1.28.30 (2021-10-05)
 ----------------------